API: Add POST method for listing apps
Add an api/v2/apps endpoint accepting a POST request containing a
JSON-formed body carrying the list of application IDs to be used as
additional filter. Since the list of supported frameworks can also grow
quite big, move that to the body, too.

With this change, the API is nearly identical to the one used by the
Ubuntu store.